Convert IllegalStateException to a ruby error on a converge failure
This commit converts an IllegalStateException to a ConfigurationError in
converge_state - the current implementation will throw when when
calling converge_result.add, as it does not know what to do with a java
Exception.
This has a knock on effect - the unexpected exception causes Logstash to
crash abruptly, regardless of whether multiple pipelines or config reload
is enabled. Additionally, the webserver is not stopped/cleaned up which
leads to a hang bug, when attempting to shutdown logstash as jruby-9.3.x
attempts to shutdown ruby threads on shutdown, but the webserver will
hang if not shut down gracefully

This commit also has some debugging info